refactor: 관리자용 예약 상세 DTO 위치 변경 (ReservationResponse -> AdminResponse)
